refactor: naming for secretsmanager interface
